+++ swarmsim-ts/src/spreadsheet/original.ts
